+++ anyblok/model/exceptions.py
@@ -1,0 +1,15 @@
+# This file is a part of the AnyBlok project
+#
+#    Copyright (C) 2017 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file,You can
+# obtain one at http://mozilla.org/MPL/2.0/.
+
+
+class ModelException(Exception):
+    """Exception for Model declaration"""
+
+
+class ViewException(ModelException):
+    """Exception for View declaration"""